@@ -12,7 +12,6 @@
     "start": "nest start",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -21,48 +20,63 @@
     "test:e2e": "jest --config ./apps/services/test/jest-e2e.json"
   },
   "dependencies": {
-    "@nestjs/axios": "^0.0.2",
-    "@nestjs/common": "^8.0.0",
-    "@nestjs/core": "^8.0.0",
+    "@elastic/elasticsearch": "~7.13.0",
+    "@nestjs/axios": "^0.0.7",
+    "@nestjs/common": "^8.4.3",
+    "@nestjs/core": "^8.4.3",
+    "@nestjs/elasticsearch": "^8.1.0",
+    "@nestjs/event-emitter": "^1.1.0",
     "@nestjs/jwt": "^8.0.0",
-    "@nestjs/mongoose": "^8.0.1",
-    "@nestjs/passport": "^8.0.1",
-    "@nestjs/platform-express": "^8.0.0",
-    "@nestjs/swagger": "^5.0.9",
-    "@types/dockerode": "^3.2.7",
+    "@nestjs/microservices": "^8.4.3",
+    "@nestjs/mongoose": "^9.0.3",
+    "@nestjs/passport": "^8.2.1",
+    "@nestjs/platform-express": "^8.4.3",
+    "@nestjs/schedule": "^1.1.0",
+    "@nestjs/swagger": "^5.2.1",
+    "@types/dockerode": "^3.3.8",
     "@types/passport-jwt": "^3.0.6",
-    "class-transformer": "^0.4.0",
-    "class-validator": "^0.13.1",
-    "dockerode": "^3.3.0",
-    "mongoose": "^6.0.3",
-    "passport": "^0.4.1",
+    "axios": "^0.26.1",
+    "bson-objectid": "^2.0.3",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
+    "dockerode": "^3.3.1",
+    "glob-to-regexp": "^0.4.1",
+    "mongoose": "^6.2.10",
+    "multer": "^1.4.4",
+    "nats": "^2.6.1",
+    "passport": "^0.5.2",
     "passport-jwt": "^4.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
-    "rxjs": "^7.2.0",
-    "swagger-ui-express": "^4.1.6"
+    "rxjs": "^7.5.5",
+    "swagger-ui-express": "^4.3.0",
+    "unzipper": "^0.10.11"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.0.0",
-    "@nestjs/schematics": "^8.0.0",
-    "@nestjs/testing": "^8.0.0",
+    "@nestjs/cli": "^8.2.5",
+    "@nestjs/schematics": "^8.0.10",
+    "@nestjs/testing": "^8.4.3",
+    "@types/cron": "^1.7.3",
     "@types/express": "^4.17.13",
-    "@types/jest": "^27.0.1",
-    "@types/node": "^16.0.0",
-    "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^4.28.2",
-    "@typescript-eslint/parser": "^4.28.2",
-    "eslint": "^7.30.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "jest": "^27.0.6",
-    "prettier": "^2.3.2",
-    "supertest": "^6.1.3",
-    "ts-jest": "^27.0.3",
-    "ts-loader": "^9.2.3",
-    "ts-node": "^10.0.0",
-    "tsconfig-paths": "^3.10.1",
-    "typescript": "^4.3.5"
+    "@types/glob-to-regexp": "^0.4.1",
+    "@types/jest": "^27.4.1",
+    "@types/multer": "^1.4.7",
+    "@types/node": "^16.11.26",
+    "@types/supertest": "^2.0.12",
+    "@types/unzipper": "^0.10.5",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "eslint": "^8.12.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.0.0",
+    "jest": "^27.5.1",
+    "prettier": "^2.6.2",
+    "supertest": "^6.2.2",
+    "ts-jest": "^27.1.4",
+    "ts-loader": "^9.2.8",
+    "ts-node": "^10.7.0",
+    "tsconfig-paths": "^3.14.1",
+    "typescript": "^4.6.3"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -86,7 +100,8 @@
     ],
     "moduleNameMapper": {
       "^@app/keycloak-auth(|/.*)$": "<rootDir>/libs/keycloak-auth/src/$1",
-      "^@app/not-found(|/.*)$": "<rootDir>/libs/not-found/src/$1"
+      "^@app/not-found(|/.*)$": "<rootDir>/libs/not-found/src/$1",
+      "^@app/event(|/.*)$": "<rootDir>/libs/event/src/$1"
     }
   }
 }
