@@ -1,63 +1,49 @@
-import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {FormsModule} from '@angular/forms';
 import {HttpClientModule} from '@angular/common/http';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
 
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {ClipboardModule} from 'ngx-clipboard';
-import {DragulaModule} from 'ng2-dragula';
+import {DndModule} from 'ngx-drag-drop';
 
 import {SharedModule} from '../shared/shared.module';
-
 import {AssignmentRoutingModule} from './assignment-routing.module';
+import {AssignmentSharedModule} from './modules/shared/shared.module';
 
-import {CreateAssignmentComponent} from './create-assignment/create-assignment.component';
-import {CreateSolutionComponent} from './create-solution/create-solution.component';
-import {MyCoursesComponent} from './my-courses/my-courses.component';
-import {SolutionComponent} from './solution/solution.component';
-import {TokenModalComponent} from './token-modal/token-modal.component';
-import {SolutionTableComponent} from './solution-table/solution-table.component';
-import {AssignmentInfoComponent} from './assignment-info/assignment-info.component';
-import {TaskListComponent} from './task-list/task-list.component';
-import {GradeFormComponent} from './grade-form/grade-form.component';
-import {CourseComponent} from './course/course.component';
-import {CreateCourseComponent} from './create-course/create-course.component';
-import {MyAssignmentsComponent} from './my-assignments/my-assignments.component';
-import {MySolutionsComponent} from './my-solutions/my-solutions.component';
-import {AuthorNameComponent} from './author-name/author-name.component';
-import {ImportExportComponent} from './import-export/import-export.component';
+import {CourseComponent} from './pages/course/course.component';
+import {CreateCourseComponent} from './pages/create-course/create-course.component';
+import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
+import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
+import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
+import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
+import {TokenModalComponent} from './pages/token-modal/token-modal.component';
+import {ConfigService} from './services/config.service';
 
 @NgModule({
-    declarations: [
-        CreateAssignmentComponent,
-        CreateSolutionComponent,
-        SolutionComponent,
-        TokenModalComponent,
-        SolutionTableComponent,
-        AssignmentInfoComponent,
-        TaskListComponent,
-        GradeFormComponent,
-        CourseComponent,
-        CreateCourseComponent,
-        MyAssignmentsComponent,
-        MySolutionsComponent,
-        MyCoursesComponent,
-        AuthorNameComponent,
-        ImportExportComponent,
-    ],
-    imports: [
-        CommonModule,
-        SharedModule,
-        FormsModule,
-        HttpClientModule,
-        NgbModule,
-        ClipboardModule,
-        DragulaModule,
-        AssignmentRoutingModule,
-    ],
-    exports: [
-        AssignmentInfoComponent,
-    ],
+  declarations: [
+    CourseComponent,
+    CreateCourseComponent,
+    CreateSolutionComponent,
+    MyAssignmentsComponent,
+    MyCoursesComponent,
+    MySolutionsComponent,
+    TokenModalComponent,
+  ],
+  imports: [
+    CommonModule,
+    SharedModule,
+    FormsModule,
+    HttpClientModule,
+    NgbModule,
+    ClipboardModule,
+    AssignmentSharedModule,
+    AssignmentRoutingModule,
+    DndModule,
+  ],
+  providers: [
+    ConfigService,
+  ],
 })
 export class AssignmentModule {
 }
