@@ -1,20 +1,32 @@
 import Task from './task';
 
 export default class Assignment {
-  _id?: string;
+  _id: string;
   token?: string;
 
   title: string;
   description: string;
-  descriptionHtml?: string;
   createdBy?: string;
   author: string;
   email: string;
-  deadline?: Date;
+  deadline?: Date | string;
+
+  classroom?: {
+    org?: string;
+    prefix?: string;
+    token?: string;
+    codeSearch?: boolean;
+  };
 
   tasks: Task[];
   solution: string;
   templateSolution: string;
 
-  static comparator = (a: Assignment, b: Assignment) => a.title.localeCompare(b.title) || a._id!.localeCompare(b._id!);
+  static comparator = (a: Assignment, b: Assignment) => a.title.localeCompare(b.title) || a._id.localeCompare(b._id);
+}
+
+export type CreateAssignmentDto = Omit<Assignment, '_id' | 'token' | 'createdBy'>;
+
+export interface UpdateAssignmentDto extends Partial<CreateAssignmentDto> {
+  token?: true;
 }
