@@ -1,4 +1,9 @@
-import TaskResult from './task-result';
+export class AuthorInfo {
+  name: string;
+  studentId: string;
+  email: string;
+  github: string;
+}
 
 export default class Solution {
   _id?: string;
@@ -6,13 +11,10 @@ export default class Solution {
   assignment: string;
 
   createdBy?: string;
-  name: string;
-  studentID: string;
-  email: string;
+  author: AuthorInfo;
   solution: string;
+  commit?: string;
 
   timestamp?: Date;
-  results?: TaskResult[];
-
-  assignee?: string;
+  points?: number;
 }
