@@ -1,20 +1,30 @@
-import {Component, Input, TemplateRef} from '@angular/core';
+import {Component, OnInit} from '@angular/core';
+import {Title} from '@angular/platform-browser';
+import {ActivatedRoute} from '@angular/router';
 import {NgbModal} from '@ng-bootstrap/ng-bootstrap';
 
 @Component({
-  selector: 'app-preview',
   templateUrl: './preview.component.html',
   styleUrls: ['./preview.component.scss'],
 })
-export class PreviewComponent {
-  @Input() title: string;
+export class PreviewComponent implements OnInit {
+  url!: string;
+  title!: string;
 
   constructor(
     private modal: NgbModal,
+    public route: ActivatedRoute,
+    private titleService: Title,
   ) {
   }
 
-  open(content: TemplateRef<any>): void {
-    this.modal.open(content, {ariaLabelledBy: 'diagram-modal-title', size: 'xl', scrollable: true});
+  ngOnInit() {
+    this.route.queryParams.subscribe(({url}) => {
+      const {pathname} = new URL(url, location.origin);
+      const fileName = pathname.substring(pathname.lastIndexOf('/') + 1);
+      this.url = url;
+      this.title = fileName;
+      this.titleService.setTitle(fileName + ' - Preview - fulib.org');
+    });
   }
 }
