@@ -2,7 +2,8 @@ export const environment = {
   production: true,
   apiURL: '/api',
   assignmentsApiUrl: '/api/v1',
-  projectsApiUrl: '/api',
+  projectsApiUrl: '/api/v1',
+  workflowsUrl: '/api/v1',
   auth: {
     url: 'https://se.uniks.de/auth',
     realm: 'fulib.org',
