@@ -2,6 +2,7 @@ import {Module} from '@nestjs/common';
 import {MongooseModule} from '@nestjs/mongoose';
 import {AssignmentModule} from '../assignment/assignment.module';
 import {AssigneeController} from './assignee.controller';
+import {AssigneeHandler} from './assignee.handler';
 import {AssigneeSchema} from './assignee.schema';
 import {AssigneeService} from './assignee.service';
 
@@ -16,7 +17,13 @@ import {AssigneeService} from './assignee.service';
     AssignmentModule,
   ],
   controllers: [AssigneeController],
-  providers: [AssigneeService],
+  providers: [
+    AssigneeService,
+    AssigneeHandler,
+  ],
+  exports: [
+    AssigneeService,
+  ],
 })
 export class AssigneeModule {
 }
