@@ -1,4 +1,6 @@
-import {ApiProperty, OmitType, PartialType, PickType} from '@nestjs/swagger';
+import {ApiProperty, ApiPropertyOptional, OmitType, PartialType, PickType} from '@nestjs/swagger';
+import {Equals, IsOptional} from 'class-validator';
+import {CreateEvaluationDto} from '../evaluation/evaluation.dto';
 import {Solution} from '../solution/solution.schema';
 import {Assignment, Task} from './assignment.schema';
 
@@ -20,12 +22,18 @@ export class UpdateAssignmentDto extends PartialType(OmitType(Assignment, [
   'token',
   'createdBy',
 ] as const)) {
+  @ApiPropertyOptional({description: 'If true, a new token is generated that replaces the old token.'})
+  @IsOptional()
+  @Equals(true)
+  token?: true;
 }
 
 export class CheckRequestDto extends PickType(Solution, ['solution'] as const) {
 }
 
-export class CheckResponseDto extends PickType(Solution, ['results'] as const) {
+export class CheckResponseDto {
+  @ApiProperty({type: [CreateEvaluationDto]})
+  results: CreateEvaluationDto[];
 }
 
 export class CheckNewRequestDto extends PickType(Assignment, ['solution', 'tasks'] as const) {
