@@ -1,10 +1,11 @@
+import {EventService} from '@app/event';
 import {UserToken} from '@app/keycloak-auth';
 import {HttpService} from '@nestjs/axios';
 import {Injectable} from '@nestjs/common';
-import {InjectConnection, InjectModel} from '@nestjs/mongoose';
-import {Connection, FilterQuery, Model} from 'mongoose';
+import {InjectModel} from '@nestjs/mongoose';
+import {FilterQuery, Model, UpdateQuery} from 'mongoose';
 import {environment} from '../environment';
-import {TaskResult} from '../solution/solution.schema';
+import {CreateEvaluationDto} from '../evaluation/evaluation.dto';
 import {generateToken, idFilter} from '../utils';
 import {CreateAssignmentDto, ReadAssignmentDto, UpdateAssignmentDto} from './assignment.dto';
 import {Assignment, AssignmentDocument, Task} from './assignment.schema';
@@ -13,15 +14,14 @@ import {Assignment, AssignmentDocument, Task} from './assignment.schema';
 export class AssignmentService {
   constructor(
     @InjectModel('assignments') private model: Model<Assignment>,
-    @InjectConnection() private connection: Connection,
     private http: HttpService,
+    private eventService: EventService,
   ) {
     this.migrate();
   }
 
   async migrate() {
-    const collection = this.connection.collection('assignments');
-    const result = await collection.updateMany({}, {
+    const result = await this.model.updateMany({}, {
       $rename: {
         userId: 'createdBy',
       },
@@ -32,11 +32,37 @@ export class AssignmentService {
     console.info('Migrated', result.modifiedCount, 'assignments');
   }
 
-  async check(solution: string, {tasks}: Pick<Assignment, 'tasks'>): Promise<TaskResult[]> {
-    return Promise.all(tasks.map(task => this.checkTask(solution, task)));
+  findTask(tasks: Task[], id: string): Task | undefined {
+    for (const task of tasks) {
+      if (task._id == id) {
+        return task;
+      }
+      const subTask = this.findTask(task.children, id);
+      if (subTask) {
+        return subTask;
+      }
+    }
+    return undefined;
   }
 
-  async checkTask(solution: string, task: Task): Promise<TaskResult> {
+  async check(solution: string, {tasks}: Pick<Assignment, 'tasks'>): Promise<CreateEvaluationDto[]> {
+    const results = await Promise.all(tasks.map(task => this.checkTasksRecursively(solution, task)));
+    return results.flatMap(x => x);
+  }
+
+  async checkTasksRecursively(solution: string, task: Task): Promise<CreateEvaluationDto[]> {
+    const [first, rest] = await Promise.all([
+      this.checkTask(solution, task),
+      Promise.all(task.children.map(t => this.checkTasksRecursively(solution, t))),
+    ]);
+    const flatRest = rest.flatMap(x => x);
+    return first ? [first, ...flatRest] : flatRest;
+  }
+
+  async checkTask(solution: string, task: Task): Promise<CreateEvaluationDto | undefined> {
+    if (!task.verification) {
+      return undefined;
+    }
     const response = await this.http.post(`${environment.compiler.apiUrl}/runcodegen`, {
       privacy: 'none',
       packageName: 'org.fulib.assignments',
@@ -44,18 +70,23 @@ export class AssignmentService {
       scenarioText: `# Solution\n\n${solution}\n\n## Verification\n\n${task.verification}\n\n`,
     }).toPromise();
     return {
-      points: response?.data.exitCode === 0 ? task.points : 0,
-      output: response?.data.output,
+      task: task._id,
+      author: 'Autograding',
+      remark: response?.data.output ?? '',
+      points: response?.data.exitCode === 0 ? Math.max(task.points, 0) : Math.min(task.points, 0),
+      snippets: [],
     };
   }
 
   async create(dto: CreateAssignmentDto, userId?: string): Promise<AssignmentDocument> {
     const token = generateToken();
-    return this.model.create({
+    const created = await this.model.create({
       ...dto,
       token,
       createdBy: userId,
     });
+    created && this.emit('created', created.id, created);
+    return created;
   }
 
   async findAll(where: FilterQuery<Assignment> = {}): Promise<ReadAssignmentDto[]> {
@@ -75,14 +106,29 @@ export class AssignmentService {
   }
 
   async update(id: string, dto: UpdateAssignmentDto): Promise<Assignment | null> {
-    return this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    const update: UpdateQuery<Assignment> = dto;
+    if (dto.token) {
+      update.token = generateToken();
+    } else {
+      delete update.token;
+    }
+    const updated = await this.model.findOneAndUpdate(idFilter(id), update, {new: true}).exec();
+    updated && this.emit('updated', id, updated);
+    return updated;
   }
 
   async remove(id: string): Promise<AssignmentDocument | null> {
-    return this.model.findOneAndDelete(idFilter(id)).exec();
+    const deleted = await this.model.findOneAndDelete(idFilter(id)).exec();
+    deleted && this.emit('deleted', id, deleted);
+    return deleted;
   }
 
   isAuthorized(assignment: Assignment, user?: UserToken, token?: string): boolean {
     return assignment.token === token || !!user && user.sub === assignment.createdBy;
+  }
+
+  private emit(event: string, id: string, assignment: Assignment) {
+    const users = [assignment.token, assignment.createdBy].filter((i): i is string => !!i);
+    this.eventService.emit(`assignment.${id}.${event}`, {event, data: assignment, users});
   }
 }
