@@ -1,7 +1,9 @@
+import {EventService} from '@app/event';
 import {UserToken} from '@app/keycloak-auth';
 import {Injectable} from '@nestjs/common';
-import {InjectConnection, InjectModel} from '@nestjs/mongoose';
-import {Connection, FilterQuery, Model} from 'mongoose';
+import {InjectModel} from '@nestjs/mongoose';
+import {FilterQuery, Model} from 'mongoose';
+import {Assignee, AssigneeDocument} from '../assignee/assignee.schema';
 import {idFilter} from '../utils';
 import {CreateCommentDto, UpdateCommentDto} from './comment.dto';
 import {Comment, CommentDocument} from './comment.schema';
@@ -9,15 +11,14 @@ import {Comment, CommentDocument} from './comment.schema';
 @Injectable()
 export class CommentService {
   constructor(
-    @InjectModel('comments') private model: Model<Comment>,
-    @InjectConnection() private connection: Connection,
+    @InjectModel('comments') public model: Model<Comment>,
+    private eventService: EventService,
   ) {
     this.migrate();
   }
 
   async migrate() {
-    const collection = this.connection.collection('comments');
-    const result = await collection.updateMany({}, {
+    const result = await this.model.updateMany({}, {
       // TODO assignment
       $rename: {
         parent: 'solution',
@@ -41,7 +42,9 @@ export class CommentService {
       distinguished,
       timestamp: new Date(),
     };
-    return this.model.create(comment);
+    const created = await this.model.create(comment);
+    this.emit('created', created);
+    return created;
   }
 
   async findAll(where: FilterQuery<Comment> = {}): Promise<CommentDocument[]> {
@@ -53,14 +56,31 @@ export class CommentService {
   }
 
   async update(id: string, dto: UpdateCommentDto): Promise<Comment | null> {
-    return this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    const updated = await this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    updated && this.emit('updated', updated);
+    return updated;
   }
 
   async remove(id: string): Promise<CommentDocument | null> {
-    return this.model.findOneAndDelete(idFilter(id)).exec();
+    const deleted = await this.model.findOneAndDelete(idFilter(id)).exec();
+    deleted && this.emit('deleted', deleted);
+    return deleted;
+  }
+
+  async removeAll(where: FilterQuery<Comment>): Promise<CommentDocument[]> {
+    const comments = await this.findAll(where);
+    await this.model.deleteMany({_id: {$in: comments.map(a => a._id)}}).exec();
+    for (let comment of comments) {
+      this.emit('deleted', comment);
+    }
+    return comments;
   }
 
   isAuthorized(comment: Comment, bearerToken: UserToken) {
     return bearerToken.sub === comment.createdBy;
+  }
+
+  private emit(event: string, comment: CommentDocument) {
+    this.eventService.emit(`comment.${comment.id}.${event}`, {event, data: comment});
   }
 }
