@@ -1,6 +1,7 @@
+import {EventService} from '@app/event';
 import {Injectable} from '@nestjs/common';
-import {InjectConnection, InjectModel} from '@nestjs/mongoose';
-import {Connection, Model} from 'mongoose';
+import {InjectModel} from '@nestjs/mongoose';
+import {Model} from 'mongoose';
 import {idFilter} from '../utils';
 import {CreateCourseDto, UpdateCourseDto} from './course.dto';
 import {Course, CourseDocument} from './course.schema';
@@ -9,14 +10,13 @@ import {Course, CourseDocument} from './course.schema';
 export class CourseService {
   constructor(
     @InjectModel('courses') private model: Model<Course>,
-    @InjectConnection() private connection: Connection,
+    private eventService: EventService,
   ) {
     this.migrate();
   }
 
   async migrate() {
-    const collection = this.connection.collection('courses');
-    const result = await collection.updateMany({}, {
+    const result = await this.model.updateMany({}, {
       $rename: {
         assignmentIds: 'assignments',
         userId: 'createdBy',
@@ -29,10 +29,12 @@ export class CourseService {
   }
 
   async create(dto: CreateCourseDto, userId?: string): Promise<CourseDocument> {
-    return this.model.create({
+    const created = await this.model.create({
       ...dto,
       createdBy: userId,
     });
+    this.emit('created', created);
+    return created;
   }
 
   async findAll(): Promise<CourseDocument[]> {
@@ -44,10 +46,18 @@ export class CourseService {
   }
 
   async update(id: string, dto: UpdateCourseDto): Promise<Course | null> {
-    return this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    const updated = await this.model.findOneAndUpdate(idFilter(id), dto, {new: true}).exec();
+    updated && this.emit('updated', updated);
+    return updated;
   }
 
   async remove(id: string): Promise<CourseDocument | null> {
-    return this.model.findOneAndDelete(idFilter(id)).exec();
+    const deleted = await this.model.findOneAndDelete(idFilter(id)).exec();
+    deleted && this.emit('deleted', deleted);
+    return deleted;
+  }
+
+  private emit(event: string, course: CourseDocument) {
+    this.eventService.emit(`comment.${course.id}.${event}`, {event, data: course});
   }
 }
