@@ -1,8 +1,11 @@
-import {Module} from '@nestjs/common';
+import {forwardRef, Module} from '@nestjs/common';
 import {MongooseModule} from '@nestjs/mongoose';
+import {AssigneeModule} from '../assignee/assignee.module';
 import {AssignmentModule} from '../assignment/assignment.module';
+import {EvaluationModule} from '../evaluation/evaluation.module';
 import {SolutionAuthGuard} from './solution-auth.guard';
 import {SolutionController} from './solution.controller';
+import {SolutionHandler} from './solution.handler';
 import {SolutionSchema} from './solution.schema';
 import {SolutionService} from './solution.service';
 
@@ -15,10 +18,13 @@ import {SolutionService} from './solution.service';
       },
     ]),
     AssignmentModule,
+    forwardRef(() => EvaluationModule),
+    AssigneeModule,
   ],
   controllers: [SolutionController],
   providers: [
     SolutionService,
+    SolutionHandler,
     SolutionAuthGuard,
   ],
   exports: [
